Introduce `--triton-to-linalg-experimental` (#114)
This pass is a combination of the following passes:

+ `--triton-to-structured`
+ `--canonicalize`
+ `--triton-arith-to-linalg`
+ `--structured-to-memref`

Once stabilized, we intend to deprecate the old `--triton-to-linalg`
pass in favour of this new pass. We expect this to happen May 1st, 2024.
At that point, we will rename the old `--triton-to-linalg` pass to
`--triton-to-linalg-legacy` and will stop future updates for it.