Use `add_triton_library` and allow optionally building the CPU backend (#120)
+ Use `add_triton_library` for all triton-shared passes and libs to be
consistent with triton. This also helps potential issues when
integrating with other backends because `add_mlir_library` exports the
targets (and therefore their dependencies) by default, but triton does
not export `TritonIR` and other passes.
+ Allow optionally build the cpu backend, this is to account for cases
where another backend just wants to use `triton-shared` as a
middle-layer and does not need the CPU backend to be installed.
+ Fix lambda capturing the context pointer by reference.
+ Add `get_codegen_implementation` to the CPUBackend, this was added
recently in triton upstream.