Add preliminary support for sign-extend ops in `PtrAnalysis` and `MaskAnalysis` (#181)
This PR adds limited support for sign-extend ops in `PtrAnalysis` and
`MaskAnalysis`.

For simplicity, we assume results of the pointer arithmetic operations
do not overflow the current platform's native bit-width. This assumption
is necessary because most of the memref and tensor related operands use
the `index` type whose width, according to the mlir documentation, is
platform specific.

Then, we simply need to compute the pointer and mask states using the
value whose sign is being extended (i.e.: the visitor of the sign-extend
op is just simply forwarding the input operand to its appropriate
visitor).

We will revisit the possibility of overflow in the future.