Support triton kernels containing early returns (#182)
Previously, the `triton-arith-to-linalg` pass always converts terminator
ops to `func.return`. With the presence of early returns, `cf.br` ops
can also appear in the IR; these `cf.br` ops are also considered
terminators and are accidentally converted to `func.return`. I have
modified the lowering logic to only convert `tt.return` ops.

A simple CPU test case is also added. As part of this test, I have
discovered that `eliminate-empty-tensors` does not support functions
with multiple return ops. Since this is mostly an optimization pass to
reduce the number of allocations, I have disabled the pass for now. We
will revisit this again if the need arises.