Support mask analysis against 0, and between two scalars (#202)
Support was added for two new scenarios:
`arith.cmpi ge %scalar, %c0`: aka offset comparison to the lower bound
of 0. Mask analysis already has an implicit assumption that the
beginning of a mask starts at 0, so support was added to allow this case
through and assumes that this comparison evaluates to true.
`arith.cmpi slt %scalar,_1 %scalar_2`: offset comparison between two
scalars. E.g.:
```
%11 = tt.expand_dims %offset
%cst_4 = arith.constant dense<324> : tensor<16x1xi64>
%23 = arith.cmpi slt, %11, %cst_4 : tensor<16x1xi64>
```
This example is notable in that we cannot take the normal approach of
computing the minimum of the lhs and rhs as the new dimension (the lhs
offset may be 0). To handle this, a ternary operator is inserted to
evaluate the comparison at runtime. If it succeeds, we keep the existing
dimensions from the lhs, otherwise we assume nothing should be
loaded/stored.

This change also adds a dump method to both `MaskState` and `PtrState`
as a small QOL improvement.