Support for Gather/Scatter access with only one non-continuous dimension (#239)
A new operation, TTS_MakeGatherScatterTensorPtrOp, was added.
It is similar to TTS_MakeTensorPtrOp, but includes gather_scatter_dim
and gather_scatter_offset to store information about the non-continuous
dimension.
With TTS_MakeIndirectTensorPtrOp, the continuous memory accesses for the
rest of the dimensions are grouped into one operation, which keeps the
high-level semantics of the operation intact.

In PtrAnalysis:
If an operation works on a single dimension, unsupported operations will
have PtrState start from the operation with the operation as the
indirect offset.
For other operations that fail the structured check, reset the dimension
using the operation as the indirect offset if it works on a single
dimension.

When dealing with mulState and addState:
For the indirect dimension, set the stride to 1 and use offset * stride
as the new offset.
If one state has a modulo and the other state is not structured, clear
the modulo and use the operand as the offset directly.

Also skipped loop case when the PtrState is not structured.

---------

Co-authored-by: Xiang Li <xiagli@microsoft.com>