Add TPtrDialect (#253)
This PR introduces TPTrDialect. This is a temporary dialect that
includes some of the ops that operate on the `!ptr.ptr` type in the
proposal here:
https://discourse.llvm.org/t/rfc-ptr-dialect-modularizing-ptr-ops-in-the-llvm-dialect/75142.
We will leverage this dialect to lower the remaining pointer operations
that cannot be handled by `triton-to-structured` and
`triton-to-unstructured` passes. Once the proposal is fully implemented,
we will migrate over to using the official ops instead of this temporary
dialect.

The code is borrowed from the author's draft implementation here with
some simplifications:
+
https://github.com/fabianmcg/mlir-address/blob/main/include/mlir/Dialect/Address/IR/AddressOps.td
+ https://github.com/llvm/llvm-project/pull/73057/files