Add option to lower triton tensor ptr ops to linalg (#254)
Add an option in `triton-arith-to-linalg` to convert triton ops on
tensor of pointers (tt.load, tt.store, tt.bitcast, tt.int_to_ptr,
tt.ptr_to_int) to linalg. This option is disabled by default and can be
enabled by setting the `convert-tensor-ptr-ops` flag to true in the
`triton-arith-to-linalg` pass.

This will help with lowering these triton ops to the equivalent ops in
the PtrDialect later.