Implement lowering of triton ops on pointer to the PtrDialect (#256)
Introduce the `triton-to-ptr` pass which lowers triton ops on pointer to
the PtrDialect. This pass only operates on triton ops on single pointers
only, and as a result, it must be run after
`triton-arith-to-linalg="tensor-ptr-to-linalg=true"`. This is our final
fallback mode for when all analyses fail, and we need to handle the
pointer arithmetic at runtime.

This pass will leave around unrealized_cast between `tt.ptr` and
`ptr.ptr`. We will have another pass that finalizes the conversion and
clean up these unrealized cast ops.