Add pass to finalize ptr cast conversion (#257)
This PR adds a new pass `reconcile-ptr-casts` to finalize the conversion
from triton pointer types to memref or the pointer type in PtrDialect.
Previous passes may have inserted unrealized_conversion_cast op between
these types, so this pass simplifies the type conversion chains and
convert the op into either `to_memref` which takes a pointer and returns
a memref, or `from_memref` which takes a memref and returns a pointer.