Support for Gather/Scatter access with only one non-continuous dimension (#266)
A new operation, TTS_MakeGatherScatterTensorPtrOp, was added.
It is similar to TTS_MakeTensorPtrOp, but includes gather_scatter_dim
and gather_scatter_offset to store information about the non-continuous
dimension.
With TTS_MakeGatherScatterTensorPtrOp, the continuous memory accesses
for the rest of the dimensions are grouped into one operation, which
keeps the high-level semantics of the operation intact.

In PtrAnalysis:
If an operation works on a single dimension, unsupported operations will
have PtrState start from the operation with the operation as the
indirect offset.
For other operations that fail the structured check, reset the dimension
using the operation as the indirect offset if it works on a single
dimension.

When dealing with mulState and addState:
For the non-continuous dimension, set the stride to 1 and use offset *
stride as the new offset.
If one state has a modulo and the other state is not structured, clear
the modulo and use the operand as the offset directly.

In StructuredToMemrefPass:
the TTS_MakeGatherScatterTensorPtrOp is lowered along with load and
store operations. These load/store operations will be expanded into copy
operations within a loop on the unstructured dimension. During each
iteration of the loop, the current row will be `reinterpret_cast to
memRefPtr[gatherOffsetElt]`.
```
// Original code snippet:
ptr = tts.make_gather_scatter_tptr(base, gatherOffsets);
load(ptr);

// Transformed code snippet:
alloc typeof(load);

foreach (offset in gatherOffsets) {
    src = reinterpret_cast(base, offset);
    dst = subview(alloc[offset]);
    copy(src, dst);
}

to_tensor(alloc);

// Original code snippet:
ptr = tts.make_gather_scatter_tptr(base, gatherOffsets);
store(stVal, ptr);

// Transformed code snippet:
foreach (offset in gatherOffsets) {
    slice = extract_slice(stVal[offset]);
    dst = reinterpret_cast(base, offset);
    materialize_in_destination(slice, dst);
}
```

---------

Co-authored-by: Xiang Li <xiagli@microsoft.com>