Support index_select style gather/scatter access. (#276)
Save stride for unstructured dimension along chain of muls.
For case like
```
row_indices[:, None] * stride_m + col_offsets[None, :] * stride_n
```
When row_indices is not continuous, the offset and row for dim 0 will be
```
offset: row_indices
stride: stride_m
```
instead of
```
offset: row_indices * stride_m
stride: 1
```

For case like
```
    offs_3d_x = offs_2d_x[:, None] // Y
    offs_3d_y = offs_2d_x[:, None] % Y

    offs_3d_z = offs_2d_y[None, :]

    offs = offs_3d_x * stride_x + offs_3d_y * stride_y + offs_3d_z * stride_z
```
When dim 0 is not continuous, the offset and stride for dim 0 will be
```
offset:  (offs_2d_x // Y * stride_x) + (offs_2d_x % Y * stride_x)
stride: 1
```
Because cannot find a stride for dim 0.


Support i64 index for TTS_MakeGatherScatterTensorPtrOp.

Fix the stride issue caused by the offset and strides of subview are
relative to the strided memref of the input memref while the offset and
stride of reinterpret_cast are relative to the base underlying memory of
the memref.

---------

Co-authored-by: Xiang Li <xiagli@microsoft.com>