Support mask on gather dim for MakeGatherScatterTensorPtrOp.
When mask is immdiate value, just update the upperbound of the gather scatter loop.
Change from
```
for i in range(0, len(index_list)):
    index = index_list[i]
    do gather/scatter on index
```
to
```
for i in range(0, min(len(index_list), mask_limit)):
    index = index_list[i]
    do gather/scatter on index
```
When mask is a dynamic value, create an scf.if to guard the gather scatter.
Change from
```
for i in range(0, len(index_list)):
    index = index_list[i]
    do gather/scatter on index
```
to
```
for i in range(0, len(index_list)):
    index = index_list[i]
    if i < mask_limit:
        do gather/scatter on index
```