Fix incorrect codegen for masks dependent on loop induction variable (#310)
Currently, for mask dependent on loop induction variable, we compute the
mask offset by adding the mask offset *before* coming into the loop by
the loop iter-arg. This is not correct when the offset has an initial
value other than 0 because then the value of the offset will always be
one iteration *after* the current iteration. This patch fixes the
codegen and adds tests for these scenarios.