update transpose injection when convert tt.reduce to linalg.reduce (#312)
when convert tt.reduce to linalg.reduce, we will insert transpose so
that the axis becomes 0. if reduce axis==0, the other dimensions can be
colllapsed. this will make the reduction easier and more efficient to
vectorize.
The pseudocode is as follows:
before 
```
... = tt.reduce axis=[x]
```
after, if x!=0
```
%v = linalg.transpose permutation [x, ...]
... = linalg.reduce %v dimensions=[0]
```

test case is `python/examples/test_reduce.py::test_reduce_max`