Adds flag for old transpose behavior on reduction conversion. (#320)
Preferably Triton Shared should not be doing any transposes for
optimizations because we lack the wider analysis to understand if these
transformations are going to be helpful or not. That will be a part of
future work.

For now, we need the old transpose behavior on the Reduce Conversion
where it just transposes rank - 1 reductions to rank - 2. This old
behavior is guarded by a added with a flag that is default to the new
behavior.