[MaskAnalysis] Fix implement of scalar conjunction in mask analysis. (#318)
The original implementation of conjunction with scalar mistakenly
assumed that the scalar argument would use it's dimension as the mask
point. This does not work because scalar arguments should be a binary
and not a pivot point. To support scalar arguments, we must either use
the mask argument of the conjunction or completely zero out the mask.
This PR changes the code generation so that select statements embedded
this pattern.