Setup clang-format action and format the whole codebase. (#349)
This PR does the following:

* Clang-Formats all C and C++ files within our repository using the LLVM
style.
* Creates a github workflow that will verify code formatting on PRs.
* Removes symlink of clang-format in favor of a permanent copy. We do
not care to keep it up to date with Triton. The likelihood of Triton
updating their clang format file is low enough for this to not be a
problem. Even if it were updated, it is not essential for us to fully
match their style. It is more important to have a permanent copy that
keeps the workflows simple.